refactor(gui): extract the pacing module and drop tiny indirections
The ipc_client's inline pacing module (with its own tests) becomes a
file; the duplicated file_url helper collapses to one copy; two
one-line single-caller wrappers in app.rs are inlined; the hotspot
geometry sorts with total_cmp instead of NaN-tolerant unwrap_or.
